Import pjproject-2.1

+++ pjmedia/src/pjmedia/types.c
@@ -1,4 +1,4 @@
-/* $Id: types.c 3715 2011-08-19 09:35:25Z nanang $ */
+/* $Id: types.c 4411 2013-03-04 04:34:38Z nanang $ */
 /*
  * Copyright (C) 2008-2011 Teluu Inc. (http://www.teluu.com)
  * Copyright (C) 2003-2008 Benny Prijono <benny@prijono.org>
@@ -37,10 +37,10 @@ PJ_DEF(const char*) pjmedia_type_name(pjmedia_type t)
 	"unknown"
     };
 
-    pj_assert(t < PJ_ARRAY_SIZE(type_names));
+    pj_assert(t < (int)PJ_ARRAY_SIZE(type_names));
     pj_assert(PJMEDIA_TYPE_UNKNOWN == 4);
 
-    if (t < PJ_ARRAY_SIZE(type_names))
+    if (t < (int)PJ_ARRAY_SIZE(type_names))
 	return type_names[t];
     else
 	return "??";
